:art: Move enum helper to support component's root
